Refactor, moved Move class inside Game class
