--- src/main/java/Move.java
@@ -1,18 +1,0 @@
-public class Move {
-
-    private final Player player;
-    private final Position position;
-
-    public Move(Player player, Position position) {
-        this.player = player;
-        this.position = position;
-    }
-
-    public Player getPlayer() {
-        return player;
-    }
-
-    public Position getPosition() {
-        return position;
-    }
-}
